requirements.txt
asgiref==3.2.10
Django==3.0.8
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
django-redis==4.12.1
djangorestframework~=3.11.0
PyJWT~=1.7.1
requests~=2.24.0
djangorestframework-jwt~=1.11.0